# packages/shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ghar Khoj - Broker-Free Room Rental Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /interactions:
    post:
      operationId: createInteraction
      tags: [interactions]
      summary: Record user interaction (view, save, rent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInteractionInput"
      responses:
        "201":
          description: Interaction recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interaction"
  /interactions/user/{userId}:
    get:
      operationId: getUserInteractions
      tags: [interactions]
      summary: Get user interactions
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [view, save, rent]
      responses:
        "200":
          description: User interactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InteractionListResponse"
...
  /contracts:
    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a rental contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractInput"
      responses:
        "201":
          description: Contract created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/match/{matchId}:
    get:
      operationId: getContractByMatch
      tags: [contracts]
      summary: Get contract for a match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/user/{userId}:
    get:
      operationId: getUserContracts
      tags: [contracts]
      summary: Get all contracts for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User contracts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractListResponse"
  /contracts/{id}/sign:
    patch:
      operationId: signContract
      tags: [contracts]
      summary: Sign a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignContractInput"
      responses:
        "200":
          description: Contract signed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}/verify:
    patch:
      operationId: verifyContract
      tags: [contracts]
      summary: Admin verify a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyContractInput"
      responses:
        "200":
          description: Contract verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
...
components:
  schemas:
...
    Interaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        roomId:
          type: integer
        type:
          type: string
          enum: [view, save, rent]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, roomId, type, createdAt]
    CreateInteractionInput:
      type: object
      properties:
        userId:
          type: integer
        roomId:
          type: integer
        type:
          type: string
          enum: [view, save, rent]
      required: [userId, roomId, type]
    InteractionListResponse:
      type: object
      properties:
        interactions:
          type: array
          items:
            $ref: "#/components/schemas/Interaction"
      required: [interactions]
...
    Contract:
      type: object
      properties:
        id:
          type: integer
        matchId:
          type: integer
        tenantId:
          type: integer
        ownerId:
          type: integer
        roomId:
          type: integer
        rentAmount:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        terms:
          type: string
        ownerSignature:
          type: string
        tenantSignature:
          type: string
        ownerSignedAt:
          type: string
          format: date-time
        tenantSignedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft, owner_signed, tenant_signed, fully_signed, verified, cancelled]
        adminVerifiedAt:
          type: string
          format: date-time
        adminNote:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, matchId, tenantId, ownerId, roomId, rentAmount, startDate, endDate, status, createdAt]
    CreateContractInput:
      type: object
      properties:
        matchId:
          type: integer
        tenantId:
          type: integer
        ownerId:
          type: integer
        roomId:
          type: integer
        rentAmount:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        terms:
          type: string
      required: [matchId, tenantId, ownerId, roomId, rentAmount, startDate, endDate]
    SignContractInput:
      type: object
      properties:
        role:
          type: string
          enum: [tenant, owner]
        signature:
          type: string
      required: [role, signature]
    VerifyContractInput:
      type: object
      properties:
        decision:
          type: string
          enum: [verified, cancelled]
        adminNote:
          type: string
      required: [decision]
    ContractListResponse:
      type: object
      properties:
        contracts:
          type: array
          items:
            $ref: "#/components/schemas/Contract"
      required: [contracts]