# packages/shared/api-zod/src/generated/types/signContractInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ghar Khoj - Broker-Free Room Rental Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SignContractInputRole } from "./signContractInputRole";

export interface SignContractInput {
  role: SignContractInputRole;
  signature: string;
}